requirements.txt
accelerate==0.34.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.1
Deprecated==1.2.14
diffusers==0.27.2
dill==0.3.8
docker-pycreds==0.4.0
fastapi==0.116.1
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gradio==5.45.0
gradio_client==1.13.0
groovy==0.1.2
h11==0.14.0
hf-xet==1.1.10
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.34.5
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
jaxtyping==0.2.34
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.43.0
markdown-it-py==2.2.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmcv-full==1.7.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==2.5.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@729261dc64c4241ea36efda84fbf532cc8b425b8
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.19.2
onnxruntime-gpu==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.7
ort-nightly-gpu==1.15.0.dev20230502003
packaging==24.1
pandas==2.2.3
peft==0.13.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==2.10.1
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pygltflib==1.16.2
Pygments==2.18.0
PyMatting==1.1.12
pymeshlab==2023.12.post2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@75ebeeaea0908c5527e7b1e305fbc7681382db47
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
ruff==0.13.0
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
semantic-version==2.10.0
sentry-sdk==2.15.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.47.3
sympy==1.13.3
termcolor==2.4.0
tifffile==2024.9.20
tokenizers==0.20.0
tomli==2.0.2
tomlkit==0.12.0
torch==2.3.1+cu121
torch_scatter==2.1.2+pt23cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.5
transformers==4.45.1
trimesh==4.4.9
triton==2.3.1
typeguard==2.13.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.31.0
wandb==0.18.3
websockets==11.0.3
wrapt==1.16.0
xformers==0.0.27
xxhash==3.5.0
yapf==0.40.2
yarl==1.13.1
zipp==3.20.2
